changed int to size_t in strlen function

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: muhahmad <muhahmad@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/05 15:26:19 by muhahmad          #+#    #+#             */
-/*   Updated: 2024/06/08 13:33:05 by muhahmad         ###   ########.fr       */
+/*   Updated: 2024/07/20 15:31:54 by muhahmad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int	length;
